# openapi.json
{"openapi":"3.1.0","info":{"title":"Yuetai Core","version":"0.1.0"},"paths":{"/api/cloud_providers/create":{"post":{"tags":["云厂商配置"],"summary":"Create Provider","operationId":"create_provider_api_cloud_providers_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloudProviderCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloudProviderOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cloud_providers/update/{record_id}":{"put":{"tags":["云厂商配置"],"summary":"Update Provider","operationId":"update_provider_api_cloud_providers_update__record_id__put","parameters":[{"name":"record_id","in":"path","required":true,"schema":{"type":"integer","minimum":1,"title":"Record Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloudProviderUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloudProviderOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cloud_providers/delete/{record_id}":{"delete":{"tags":["云厂商配置"],"summary":"Delete Provider","operationId":"delete_provider_api_cloud_providers_delete__record_id__delete","parameters":[{"name":"record_id","in":"path","required":true,"schema":{"type":"integer","minimum":1,"title":"Record Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cloud_providers/pageList":{"get":{"tags":["云厂商配置"],"summary":"List Providers","operationId":"list_providers_api_cloud_providers_pageList_get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":10,"title":"Page Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloudProviderPage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cloud_regions/list":{"get":{"tags":["云区域配置"],"summary":"List Regions","operationId":"list_regions_api_cloud_regions_list_get","parameters":[{"name":"provider_code","in":"query","required":true,"schema":{"type":"string","title":"Provider Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloudRegionOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cloud_zones/list":{"get":{"tags":["云可用区"],"summary":"List Zones","operationId":"list_zones_api_cloud_zones_list_get","parameters":[{"name":"provider_code","in":"query","required":false,"schema":{"type":"string","default":"aliyun","title":"Provider Code"}},{"name":"region_id","in":"query","required":false,"schema":{"type":"string","default":"cn-qingdao","title":"Region Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloudZoneOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cloud_certificates/create":{"post":{"tags":["云凭证"],"summary":"Create Certificate","operationId":"create_certificate_api_cloud_certificates_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloudCertificateCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloudCertificateOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cloud_certificates/pageList":{"get":{"tags":["云凭证"],"summary":"List Certificates","operationId":"list_certificates_api_cloud_certificates_pageList_get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":10,"title":"Page Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloudCertificatePage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cloud_certificates/update/{record_id}":{"put":{"tags":["云凭证"],"summary":"Update Certificate","operationId":"update_certificate_api_cloud_certificates_update__record_id__put","parameters":[{"name":"record_id","in":"path","required":true,"schema":{"type":"integer","minimum":1,"title":"Record Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloudCertificateUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloudCertificateOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cloud_certificates/delete/{record_id}":{"delete":{"tags":["云凭证"],"summary":"Delete Certificate","operationId":"delete_certificate_api_cloud_certificates_delete__record_id__delete","parameters":[{"name":"record_id","in":"path","required":true,"schema":{"type":"integer","minimum":1,"title":"Record Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/resource_groups/create":{"post":{"tags":["资源组管理"],"summary":"Create Group","operationId":"create_group_api_resource_groups_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceGroupCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceGroupOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/resource_groups/update/{record_id}":{"put":{"tags":["资源组管理"],"summary":"Update Group","operationId":"update_group_api_resource_groups_update__record_id__put","parameters":[{"name":"record_id","in":"path","required":true,"schema":{"type":"integer","minimum":1,"description":"记录ID","title":"Record Id"},"description":"记录ID"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceGroupUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceGroupOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/resource_groups/delete/{record_id}":{"delete":{"tags":["资源组管理"],"summary":"Delete Group","operationId":"delete_group_api_resource_groups_delete__record_id__delete","parameters":[{"name":"record_id","in":"path","required":true,"schema":{"type":"integer","minimum":1,"description":"记录ID","title":"Record Id"},"description":"记录ID"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/resource_groups/pageList":{"get":{"tags":["资源组管理"],"summary":"List Groups","operationId":"list_groups_api_resource_groups_pageList_get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"页码（从1开始）","default":1,"title":"Page"},"description":"页码（从1开始）"},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"每页条数","default":10,"title":"Page Size"},"description":"每页条数"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceGroupPage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/resource_group_bindings/bind":{"post":{"tags":["资源组绑定"],"summary":"Bind Resource","operationId":"bind_resource_api_resource_group_bindings_bind_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceGroupBindingCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceGroupBindingOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/resource_group_bindings/{binding_id}":{"delete":{"tags":["资源组绑定"],"summary":"Unbind Resource","operationId":"unbind_resource_api_resource_group_bindings__binding_id__delete","parameters":[{"name":"binding_id","in":"path","required":true,"schema":{"type":"integer","description":"绑定记录 ID","title":"Binding Id"},"description":"绑定记录 ID"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/resource_group_bindings/group/{group_id}/page":{"get":{"tags":["资源组绑定"],"summary":"List Bindings Page","operationId":"list_bindings_page_api_resource_group_bindings_group__group_id__page_get","parameters":[{"name":"group_id","in":"path","required":true,"schema":{"type":"integer","title":"Group Id"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":10,"title":"Page Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceGroupBindingPage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/vpc/list":{"get":{"tags":["vpc"],"summary":"List Vpcs","operationId":"list_vpcs_api_vpc_list_get","parameters":[{"name":"provider_code","in":"query","required":false,"schema":{"type":"string","default":"aliyun","title":"Provider Code"}},{"name":"region_id","in":"query","required":false,"schema":{"type":"string","default":"cn-qingdao","title":"Region Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VpcOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/vpc/pageList":{"get":{"tags":["vpc"],"summary":"List Page","operationId":"list_page_api_vpc_pageList_get","parameters":[{"name":"provider_code","in":"query","required":false,"schema":{"type":"string","default":"aliyun","title":"Provider Code"}},{"name":"region_id","in":"query","required":false,"schema":{"type":"string","default":"cn-qingdao","title":"Region Id"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Page"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Page Size"}},{"name":"total","in":"query","required":true,"schema":{"type":"integer","title":"Total"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VpcOut"},"title":"Items"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VpcPage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"CloudCertificateCreate":{"properties":{"cloud_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cloud Code","description":"云凭证唯一code"},"cloud_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cloud Name","description":"云凭证名称"},"cloud_access_key_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cloud Access Key Id","description":"AccessKey ID"},"cloud_access_key_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cloud Access Key Secret","description":"AccessKey Secret（加密存储）"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"描述信息"}},"type":"object","title":"CloudCertificateCreate"},"CloudCertificateOut":{"properties":{"id":{"type":"integer","title":"Id"},"cloud_code":{"type":"string","title":"Cloud Code"},"cloud_name":{"type":"string","title":"Cloud Name"},"cloud_access_key_id":{"type":"string","title":"Cloud Access Key Id"},"cloud_access_key_secret":{"type":"string","title":"Cloud Access Key Secret"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["id","cloud_code","cloud_name","cloud_access_key_id","cloud_access_key_secret","description"],"title":"CloudCertificateOut"},"CloudCertificatePage":{"properties":{"page":{"type":"integer","title":"Page"},"pageSize":{"type":"integer","title":"Pagesize"},"total":{"type":"integer","title":"Total"},"items":{"items":{"$ref":"#/components/schemas/CloudCertificateOut"},"type":"array","title":"Items"}},"type":"object","required":["page","pageSize","total","items"],"title":"CloudCertificatePage"},"CloudCertificateUpdate":{"properties":{"cloud_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cloud Name","description":"云凭证名称"},"cloud_access_key_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cloud Access Key Id","description":"AccessKey ID"},"cloud_access_key_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cloud Access Key Secret","description":"AccessKey Secret（加密存储）"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"描述信息"}},"type":"object","title":"CloudCertificateUpdate"},"CloudProviderCreate":{"properties":{"provider_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Provider Code","description":"云厂商唯一code"},"provider_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Provider Name","description":"云厂商名称"},"access_key_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Access Key Id","description":"AccessKey ID"},"access_key_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Access Key Secret","description":"AccessKey Secret（加密存储）"},"endpoint":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Endpoint","description":"默认 Endpoint"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"描述信息"}},"type":"object","title":"CloudProviderCreate"},"CloudProviderOut":{"properties":{"id":{"type":"integer","title":"Id"},"provider_code":{"type":"string","title":"Provider Code"},"provider_name":{"type":"string","title":"Provider Name"},"access_key_id":{"type":"string","title":"Access Key Id"},"access_key_secret":{"type":"string","title":"Access Key Secret"},"endpoint":{"type":"string","title":"Endpoint"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["id","provider_code","provider_name","access_key_id","access_key_secret","endpoint","description"],"title":"CloudProviderOut"},"CloudProviderPage":{"properties":{"page":{"type":"integer","title":"Page"},"pageSize":{"type":"integer","title":"Pagesize"},"total":{"type":"integer","title":"Total"},"items":{"items":{"$ref":"#/components/schemas/CloudProviderOut"},"type":"array","title":"Items"}},"type":"object","required":["page","pageSize","total","items"],"title":"CloudProviderPage"},"CloudProviderUpdate":{"properties":{"provider_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Provider Name","description":"云厂商名称"},"access_key_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Access Key Id","description":"AccessKey ID"},"access_key_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Access Key Secret","description":"AccessKey Secret（加密存储）"},"endpoint":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Endpoint","description":"默认 Endpoint"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"描述信息"}},"type":"object","title":"CloudProviderUpdate"},"CloudRegionOut":{"properties":{"provider_code":{"type":"string","title":"Provider Code"},"region_id":{"type":"string","title":"Region Id"},"region_name":{"type":"string","title":"Region Name"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","created_at","updated_at"],"title":"CloudRegionOut"},"CloudZoneOut":{"properties":{"id":{"type":"integer","title":"Id"},"provider_code":{"type":"string","title":"Provider Code"},"region_id":{"type":"string","title":"Region Id"},"zone_id":{"type":"string","title":"Zone Id"},"zone_name":{"type":"string","title":"Zone Name"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","provider_code","region_id","zone_id","zone_name","created_at","updated_at"],"title":"CloudZoneOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ResourceGroupBindingCreate":{"properties":{"resource_group_id":{"type":"integer","title":"Resource Group Id","description":"资源组 ID"},"resource_type":{"type":"string","title":"Resource Type","description":"资源类型"},"resource_id":{"type":"string","title":"Resource Id","description":"资源唯一 ID"}},"type":"object","required":["resource_group_id","resource_type","resource_id"],"title":"ResourceGroupBindingCreate"},"ResourceGroupBindingOut":{"properties":{"resource_group_id":{"type":"integer","title":"Resource Group Id"},"resource_type":{"type":"string","title":"Resource Type"},"resource_id":{"type":"string","title":"Resource Id"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["resource_group_id","resource_type","resource_id","id","created_at"],"title":"ResourceGroupBindingOut"},"ResourceGroupBindingPage":{"properties":{"page":{"type":"integer","title":"Page"},"pageSize":{"type":"integer","title":"Pagesize"},"total":{"type":"integer","title":"Total"},"items":{"items":{"$ref":"#/components/schemas/ResourceGroupBindingOut"},"type":"array","title":"Items"}},"type":"object","required":["page","pageSize","total","items"],"title":"ResourceGroupBindingPage"},"ResourceGroupCreate":{"properties":{"name":{"type":"string","title":"Name"},"code":{"type":"string","title":"Code"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["name","code"],"title":"ResourceGroupCreate"},"ResourceGroupOut":{"properties":{"name":{"type":"string","title":"Name"},"code":{"type":"string","title":"Code"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["name","code","id"],"title":"ResourceGroupOut"},"ResourceGroupPage":{"properties":{"page":{"type":"integer","title":"Page"},"pageSize":{"type":"integer","title":"Pagesize"},"total":{"type":"integer","title":"Total"},"items":{"items":{"$ref":"#/components/schemas/ResourceGroupOut"},"type":"array","title":"Items"}},"type":"object","required":["page","pageSize","total","items"],"title":"ResourceGroupPage"},"ResourceGroupUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","title":"ResourceGroupUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VpcOut":{"properties":{"name":{"type":"string","title":"Name","description":"VPC 名称"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"VPC 描述信息"},"resource_group_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Resource Group Id","description":"资源组ID"},"cloud_provider_id":{"type":"integer","title":"Cloud Provider Id","description":"云厂商ID"},"cloud_certificate_id":{"type":"integer","title":"Cloud Certificate Id","description":"云凭证ID"},"region_id":{"type":"string","title":"Region Id","description":"区域ID"},"network_type":{"type":"string","title":"Network Type","description":"网络类型，例如 VPC/CLASSIC"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["name","cloud_provider_id","cloud_certificate_id","region_id","network_type","id","created_at","updated_at"],"title":"VpcOut"},"VpcPage":{"properties":{"total":{"type":"integer","title":"Total"},"items":{"items":{"$ref":"#/components/schemas/VpcOut"},"type":"array","title":"Items"}},"type":"object","required":["total","items"],"title":"VpcPage"}}}}